backend/requirements.txt
fastapi==0.115.0
mangum==0.17.0
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
uvicorn==0.32.0
